third_party/blink/web_tests/platform/generic/fast/css/content-language-no-content-expected.txt

Test for bug 76701: map HTTP-EQUIV content-language to -webkit-locale. This particular test tests that a content-language with missing content is effectively ignored. This expectation may change, see comments in bug.

PASS languageOfNode('x') is "auto"
PASS languageOfNode('y') is "\"ar\""
PASS successfullyParsed is true

TEST COMPLETE
